pyproject.toml
[tool.poetry]
name = "date"
version = "0.1.0"
description = ""
authors = ["altescy <altescy@fastmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
xallennlp = {git = "https://github.com/altescy/xallennlp", rev = "v0.1.0"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
python-language-server = "^0.36.2"
pysen = {version = "^0.9.1", extras = ["lint"]}

[tool.pysen]
version = "0.9"

[tool.pysen-cli]
settings_dir = "."

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "strict"
line_length = 88
py_version = "py38"
[[tool.pysen.lint.mypy_targets]]
  paths = ["."]

[tool.pysen.lint.source]
  includes = ["."]
  excludes = [".venv/"]

[tool.pytest.ini_options]
testpaths = ["tests/"]
log_format = "%(asctime)s - %(levelname)s - %(name)s - %(message)s"
log_level = "DEBUG"
markers = ["java"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]


[tool.black] # automatically generated by pysen
line-length = 88
target-version = ["py38"]

[tool.isort] # automatically generated by pysen
default_section = "THIRDPARTY"
ensure_newline_before_comments = true
force_grid_wrap = 0
force_single_line = false
include_trailing_comma = true
line_length = 88
multi_line_output = 3
use_parentheses = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
